[#85450732] decorate crowdblog posts controller
